Split the alg into scramble + solution.
Turns out using the timestamps to do this is _not_ robust
as the first move of the actual solve can sometimes get a
zero timestamp, plus the coalescing makes the code pretty
messy. Set this aside with the intent of recording the
split at solve time and storing it in the data.