Rotate the solution to look more natural on playback.
Use the analyzer's notion of where the left block belongs to
display the whole solution in a different (hopefully more
natural) orientation.

Surely there are bugs in the repetitive edits to Analyzer.ts
but they don't yet appear as we're not showing the user the
shorter version of the solutions yet (and this also was not
working before).