Tidy up GAN bruteforcing flow.
Ensure that the count of moves tracks what the user has done,
that the cube is reset to a pristine state after decryption,
and that there's hopefully enough debug info on the screen that
can be used in case of failure.

Overall the scheme seems to work well enough to say that the v2
protocol is usable for normal end users, though the wait is rather
longer than one might wish.