Fix an incorrect orientation pushBack.
To my great surprise, orientation wasn't being inverted correctly
in one spot which doesn't seem to affect most reconstructions but
did affect one for me today.

On the way, got distracted removing `premove` because it too would
cause bugs if it was set on later stages and not transformed by
pushBacks resulting from wide move introductions; and it just seems
to make a lot of the code simpler to put the premove right onto the
start of its solution stage than to have a special move for every
stage. This somewhat obscures the actual fix in this patch, though
it too is a legitimate fix there is no test that validates it.