Add support for blanks.
The words code didn't look to play blanks when a blank was indicated
as a capital letter. Fix it so that it will do so, and add a basic
test for that.

While I was at it, notice that the currentPlayer fix broke another
test and fix it to have Bob play the second move.