Deal out ingredients at game setup time.
The forest ingredients weren't being dealt but should be seen
before the player is forced to make any decisions, so deal them
out at setup time. As a result, label all the spaces on the board
and fix up card sizing to be based on the container instead of
hardcoded into the cards themselves, and fix tests to account for
the fact that cards are now already drawn from the ingredient deck.

While we're in there, show the facedown draw piles and their counts.