Finally get turn order right.
Put players in forward turn order in the list based on flask placement
and deal with the backwardsness directly in the cube placement code
instead. Use the new turn order in combination with cube placement
to correctly assess who is being waited for at any given moment,
and remove the end of forage commit expectation so that foraging
from the deck can proceed out of order while choosing specific cards
causes waiting.

Start to think about end of round detection, after which it should
be possible to implement all action phases and complete an actual
game.