Somewhat complex end of round logic.
Embed in the client the ability to have all players automatically
'pass' at the end of the round so that the round end is detected
by everyone passing and the last passing player can set up the game
for the next round.

This somewhat messy solution does enable everyone to play out of
order as far as possible and advances cleanly between rounds.

One potential issue is that the round number isn't updated until
it is time to place cubes and it seems like maybe that should be
end of round book keeping instead, possibly with a special case
for round 0 -> 1.