Shopping, testing, drinking, and herbalist.
Fixed a bug where the herbalist was only giving the user two
ingredients instead of three.

Fixed a bug where round increment was in the wrong spot.

Made it possible to renounce actions, test on a student, and
drink potions.

Made it possible to buy level I artifacts and pay for them,
though they have no game effect.